% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{abundance_matrix}
\alias{abundance_matrix}
\title{Return an abundance matrix}
\usage{
abundance_matrix(ta, sample_name = sample, taxon_name = taxon, var = abundance)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{sample_name}{The name of the variable in the sample table to use as row
names (unquoted).}

\item{taxon_name}{The name of the variable in the taxon table to use as
column names (unquoted).}
}
\value{
A matrix with abundance values.
}
\description{
This function returns a matrix with taxon counts; the rows are samples and
the column are taxa.
}
