% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{abundances_matrix}
\alias{abundances_matrix}
\title{Return an abundances matrix}
\usage{
abundances_matrix(
  ta,
  value = abundance,
  sample_name = sample,
  taxon_name = taxon
)
}
\arguments{
\item{ta}{A tidyamplicons object.}

\item{value}{The name of a variable in the abundances table that contains the
abundances (unquoted). Could be relative abundances, if present.}

\item{sample_name}{The name of the variable in the sample table to use as row
names (unquoted).}

\item{taxon_name}{The name of the variable in the taxon table to use as
column names (unquoted).}
}
\value{
A matrix with abundance values.
}
\description{
This function returns the abundances table in the form of a matrix where the
rows are samples and the column are taxa.
}
