% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_alphas}
\alias{add_alphas}
\title{Add alpha diversity measures}
\usage{
add_alphas(ta)
}
\arguments{
\item{ta}{tidytacos object.}
}
\description{
\code{add_alphas} adds two alpha diversity measures to the
samples tibble of a tidytacos object.
}
\details{
This function adds two alpha diversity measures (observed and inverse
Simpson) to the samples tibble of a tidytacos object under the variable
names observed and inverse_simpson, respectively. This function will also
add relative abundances if not present using \code{\link{add_rel_abundance}}.
}
\examples{
# Initiate abundance matrix
x <- matrix(
 c(1500, 1300, 280, 356),
 ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
                     taxa_are_columns = FALSE
                     )

# Add total abundance
data <- data \%>\%
 add_alphas()
}
