% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_cluster}
\alias{add_cluster}
\title{Add cluster number}
\usage{
add_cluster(ta, n_clusters)
}
\arguments{
\item{ta}{Tidyamplicons object.}

\item{n_clusters}{Numerical. Number of desired clusters.}
}
\description{
\code{add_cluster} adds a new variable to the samples tibble of a
tidyamplicons object defining to what cluster a sample belongs.
}
\details{
This function calculates the Bray-Curtis distance between samples followed by
hierarchical average linkage clustering of samples. The user provides a
number of desired clusters which will be used to assign the samples to. A new
variable named "cluster" will be added to the samples tibble of a
tidyamplicons object defining to what cluster a sample belongs.
}
\examples{
# Initiate abundance matrix
x <- matrix(
 c(1500, 1300, 280, 356),
 ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidyamplicons object
data <- create_tidyamplicons(x,
                     taxa_are_columns = FALSE
                     )

# Add total abundance
data <- data \%>\%
 add_cluster(n_clusters = 5)

}
