% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codifab.R
\name{add_copca}
\alias{add_copca}
\title{Add compositional principal components to the sample table}
\usage{
add_copca(ta)
}
\arguments{
\item{ta}{a tidytacos object.}
}
\description{
Add compositional principal components to the sample table
}
