% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_taxa.R
\name{add_jervis_bardy}
\alias{add_jervis_bardy}
\title{Apply the taxon QC method of Jervis-Bardy}
\usage{
add_jervis_bardy(ta, dna_conc, sample_condition = T, min_pres = 3)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{dna_conc}{A variable in the samples table that contains dna
concetrations (unquoted).}

\item{sample_condition}{An optional extra condition that samples must pass
before calculations.}

\item{min_pres}{The minimum number of samples a taxon has to be present in
for its correlation to be calculated.}
}
\description{
Function to estimate spearman correlation between relative abundance and
sample dna concentration, for each taxon.
}
\details{
See:
J. Jervis-Bardy et al., “Deriving accurate microbiota profiles from
human samples with low bacterial content through post-sequencing processing
of Illumina MiSeq data,” Microbiome, vol. 3, no. 1, Art. no. 1, 2015, doi:
10.1186/s40168-015-0083-8.
}
