% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codifab.R
\name{add_logratios}
\alias{add_logratios}
\title{Add logratios}
\usage{
add_logratios(ta, max_taxa = 30)
}
\arguments{
\item{ta}{A tidyamplicons object}

\item{max_taxa}{The maximum number of taxa to use}
}
\value{
A tidyamplicons object with an extra table logratios
}
\description{
This function computes pairwise logratio values between all taxa and adds
these to the tidyamplicons object in the form of a table called logratios.
}
\details{
If max_taxa is greater than the number of taxa, the taxa with the highest
occurrence will be selected.

IMPORTANT: this function add pseudocounts of one to all abundances before
calculating the logratios.
}
