% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_taxa.R
\name{add_max_rel_abundance}
\alias{add_max_rel_abundance}
\title{Add the maximum relative abundance of taxa to the taxon table}
\usage{
add_max_rel_abundance(ta)
}
\arguments{
\item{ta}{a tidytacos object}
}
\description{
Add the maximum relative abundance of taxa to the taxon table
}
