% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_taxa.R
\name{add_mean_rel_abundances}
\alias{add_mean_rel_abundances}
\title{Add average relative abundances}
\usage{
add_mean_rel_abundances(ta, condition = NULL, test = NULL)
}
\arguments{
\item{ta}{A tidyamplicons object}

\item{condition}{A condition variable (character)}

\item{test}{Differential abundance test to perform}
}
\value{
A tidyamplicons object
}
\description{
This function adds mean relative abundance values for each taxon to the taxa
table, overall or per sample group.
}
\details{
If `condition` is specified, the mean relative abundances will be calculated
separately for each group defined by the condition variable. This variable
should be present in the sample table.

If `condition` is specified, differential abundance testing can be performed
by setting the `test` argument. Options are NULL (default), "wilcox" or
"t-test".
}
