% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_taxa.R
\name{add_occurrences}
\alias{add_occurrences}
\title{Add taxon occurrences to the taxon table}
\usage{
add_occurrences(ta, condition = NULL, relative = F, fischer_test = F)
}
\description{
Adds taxon occurrences (overall or per condition) to the taxa table.
}
\details{
Condition should be a categorical variable present in the samples table.
Supply condition as a string.
}
