% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_pcoa}
\alias{add_pcoa}
\title{Add first two dimensions of PCOA}
\usage{
add_pcoa(ta)
}
\arguments{
\item{ta}{tidytacos object.}
}
\description{
\code{add_pcoa} adds the first two dimensions of a principal components
analysis on a Bray-Curtis dissimilarity matrix to two new variables of the
samples tibble of a tidytacos object.
}
\details{
This function calculates the Bray-Curtis distance between samples followed by
a principal components analysis. It will then add the two first dimensions to
the samples tibble of a tidytacos object named "pcoa1" and "pcoa2". This
function will also add relative abundances if not present using
\code{\link{add_rel_abundance}}.
}
\examples{
# Initiate abundance matrix
x <- matrix(
 c(1500, 1300, 280, 356, 456, 678),
 ncol = 3
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2", "sample3")

# Convert to tidytacos object
data <- create_tidytacos(x,
                     taxa_are_columns = FALSE
                     )

# Add total abundance
data <- data \%>\%
 add_pcoa()

}
