% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_taxa.R
\name{add_presence_counts}
\alias{add_presence_counts}
\title{Add absolute occurrences of taxa to the taxon table}
\usage{
add_presence_counts(ta, condition)
}
\description{
Adds taxon presence and absence counts in sample conditions to the taxa
table, as well as a fisher exact test for differential presence.
}
\details{
Condition is a variable that should be present in the samples table.

DEPRECATED, use \code{\link{add_occurrences}}
}
