% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_taxa.R
\name{add_rel_occurrence}
\alias{add_rel_occurrence}
\title{Add the relative occurrence of taxa to the taxon table}
\usage{
add_rel_occurrence(ta)
}
\description{
DEPRECATED, use \code{\link{add_occurrences}}
}
\details{
Credits to Wenke Smets for the idea and initial implementation.
}
