% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_sample_tibble}
\alias{add_sample_tibble}
\title{Add sample table to the tidytacos object}
\usage{
add_sample_tibble(ta, sample_tibble)
}
\arguments{
\item{ta}{tidytacos object.}

\item{sample_tibble}{A tibble containing sample data for each sample. samples
should be rows, while sample data should be columns. At least one column
name needs to be shared with the sample tibble of ta. The default shared
column name is 'sample'.}
}
\description{
\code{add_sample_tibble} adds a sample tibble to the tidytacos object.
}
\details{
This function adds a sample tibble containing metadata for each sample to the
tidytacos object. It is used after initiating a tidytacos object
using a numerical abundance matrix and the function
\code{\link{create_tidytacos}}. Also see \code{\link{add_taxon_tibble}}
to update the taxon data of the tidytacos object.
}
\examples{
# Initiate abundance matrix
x <- matrix(
 c(1500, 1300, 280, 356),
 ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
                     taxa_are_columns = FALSE
                     )

# Initiate sample tibble
sample <- c("sample1", "sample2")
environment <- c("food fermentation", "human stool")
sample_tibble <- tibble::tibble(sample, environment)

# Add sample tibble to tidytacos object
data <- data \%>\%
add_sample_tibble(sample_tibble)

}
