% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{add_sparcc_network_clusters}
\alias{add_sparcc_network_clusters}
\title{Performs SparCC network analysis on a tidytacos object 
and then performs Markov Clustering on the network to 
annotate taxa of the largest clusters in the tidytacos object.}
\usage{
add_sparcc_network_clusters(
  ta,
  rarefact = 0.05,
  network_tresh = 0.1,
  min_n_cluster = 3,
  taxon_name = taxon,
  sample_name = sample
)
}
\arguments{
\item{ta}{a tidytacos object.}

\item{rarefact}{Rarefaction degree: percentage of samples the taxon needs to be present in.}

\item{network_tresh}{absolute value of correlations below this threshold are 
filtered out.}

\item{min_n_cluster}{minimum number of taxa per cluster, smaller clusters are filtered out.}

\item{taxon_name}{unique name to use for the taxa, by default taxon_id is used.}

\item{sample_name}{unique name to use for the samples, by default sample_id is used.}
}
\description{
Performs SparCC network analysis on a tidytacos object 
and then performs Markov Clustering on the network to 
annotate taxa of the largest clusters in the tidytacos object.
}
