% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_taxa.R
\name{add_taxon_name}
\alias{add_taxon_name}
\title{Create sensible names for the taxa and add to taxon table}
\usage{
add_taxon_name(ta, method = "mean_rel_abundance", include_species = F)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{method}{The method by which to arrange the taxon names. Currently only
mean_rel_abundance.}

\item{include_species}{Whether to include the species name or not.}
}
\description{
Create sensible names for the taxa and add to taxon table
}
