% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_taxa.R
\name{add_taxon_name_color}
\alias{add_taxon_name_color}
\title{Create taxon names suitable for visualization with color}
\usage{
add_taxon_name_color(
  ta,
  method = "total_rel_abundance",
  n = 12,
  samples = NULL,
  taxa = NULL
)
}
\description{
Create taxon names suitable for visualization with color
}
