% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{aggregate_samples}
\alias{aggregate_samples}
\title{Aggregate samples with identical values for all metadata}
\usage{
aggregate_samples(ta)
}
\arguments{
\item{ta}{A tidytacos object.}
}
\description{
Aggregate samples with identical values for all metadata
}
