% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{aggregate_taxa}
\alias{aggregate_taxa}
\title{Aggregate taxa on a given taxonomic rank}
\usage{
aggregate_taxa(ta, rank = NULL)
}
\description{
There are two ways to call this function:
}
\details{
* If the rank you are interested in is in the standard list, just supply it
as an argument.
* If not, delete all taxon variables except taxon_id and the ranks you are
still interested in prior to calling this function.
}
