% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{as_abundances}
\alias{as_abundances}
\title{Convert matrix with abundances to tidy data frame}
\usage{
as_abundances(abundances_matrix, taxa_are_columns = TRUE,
  value = "abundance")
}
\arguments{
\item{abundances_matrix}{The ambundance matrix that will be converted.}

\item{taxa_are_columns}{A logical scalar. Are the taxa defined in columns?
Default is TRUE.}

\item{value}{Name of resulting colum containing the abundance data. Default
is "abundance".}
}
\description{
\code{as_abundances} returns a tidy data frame given a numerical abundance
matrix.
}
\details{
This function will convert a numerical abundance matrix into a tidy data
frame. To convert a tidy data frame into a numerical abundance matrix
use \code{\link{as_abundances_matrix}}.
}
