% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{as_abundances_matrix}
\alias{as_abundances_matrix}
\title{Convert abundances tidy data frame to matrix.}
\usage{
as_abundances_matrix(abundances, value = abundance)
}
\arguments{
\item{abundances}{The abundance tidy data frame that will be converted.}

\item{value}{Name of colum containing the abundance data. Default is
"abundance".}
}
\description{
\code{as_abundances_matrix} returns a numerical matrix given a tidy
abundances data frame.
}
\details{
This function will convert a abundances tidy data frame into a numerlical
abundance matrix. To convert a numerical abundance matrix into a abundances
tidy data frame use \code{\link{as_abundances_matrix}}.
}
