% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{as_counts}
\alias{as_counts}
\title{Convert matrix with counts to tidy data frame}
\usage{
as_counts(counts_matrix, taxa_are_columns = TRUE, value = "counts")
}
\arguments{
\item{counts_matrix}{The ambundance matrix that will be converted.}

\item{taxa_are_columns}{A logical scalar. Are the taxa defined in columns?
Default is TRUE.}

\item{value}{Name of resulting colum containing the count data. Default
is "counts".}
}
\description{
\code{as_counts} returns a tidy data frame given a numerical counts
matrix.
}
\details{
This function will convert a numerical counts matrix into a tidy data
frame. To convert a tidy data frame into a numerical counts matrix
use \code{\link{as_counts_matrix}}.
}
