% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{as_counts_matrix}
\alias{as_counts_matrix}
\title{Convert counts tidy data frame to matrix.}
\usage{
as_counts_matrix(counts, value = count)
}
\arguments{
\item{counts}{The counts tidy data frame that will be converted.}

\item{value}{Name of colum containing the counts data. Default is
"counts".}
}
\description{
\code{as_counts_matrix} returns a numerical matrix given a tidy
counts data frame.
}
\details{
This function will convert a counts tidy data frame into a numerlical
counts matrix. To convert a numerical counts matrix into a counts
tidy data frame use \code{\link{as_counts_matrix}}.
}
