% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{as_tidytacos}
\alias{as_tidytacos}
\title{Convert phyloseq object to tidytacos object}
\usage{
as_tidytacos(ps)
}
\arguments{
\item{ps}{Phyloseq object.}
}
\description{
\code{as_tidytacos} returns a tidytacos object given a phyloseq
object.
}
\details{
This function will convert a phyloseq object into a tidytacos object. To
convert from a tidytacos object to a phyloseq object use
\code{\link{as_phyloseq}}.
}
