% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{change_id_samples}
\alias{change_id_samples}
\title{Change sample IDs to a given expression}
\usage{
change_id_samples(ta, sample_id_new)
}
\arguments{
\item{ta}{A tidytacos object}

\item{sample_id_new}{An expression that evaluates to a unique sample
identifier}
}
\description{
Change sample IDs to a given expression
}
