% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{change_id_taxa}
\alias{change_id_taxa}
\title{Change taxon IDs to a given expression}
\usage{
change_id_taxa(ta, taxon_id_new)
}
\arguments{
\item{ta}{A tidytacos object}

\item{taxon_id_new}{An expression that evaluates to a unique taxon
identifier}
}
\description{
Change taxon IDs to a given expression
}
