% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_test_ta}
\alias{create_test_ta}
\title{Create a tidytacos object for testing/example purporses}
\usage{
create_test_ta()
}
\value{
a small tidytacos object
}
\description{
Create a tidytacos object for testing/example purporses
}
