% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{get_history_plot}
\alias{get_history_plot}
\title{Return a history plot of the samples}
\usage{
get_history_plot(ta, col = NULL)
}
\description{
DEPRECATED, this function is kept for historical reasons and will probably
not work
}
