% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_numbers}
\alias{get_numbers}
\title{Get numbers}
\usage{
get_numbers(ta)
}
\arguments{
\item{ta}{Tidyamplicons object.}
}
\description{
\code{get_numbers} returns the number of samples, taxa and reads in a
tidyamplicons object.
}
\details{
This function returns the number of samples, taxa and reads in a tidyamplicons
object, stored in a named numeric vector. To print the numbers, use
\code{\link{report_numbers}} instead.
}
\examples{
# Initiate abundance matrix
x <- matrix(
 c(1500, 1300, 280, 356),
 ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidyamplicons object
data <- create_tidyamplicons(x,
                     taxa_are_columns = FALSE
                     )
# Report numbers
numbers <- data \%>\%
 get_numbers()
}
