% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{list_taxa_per_condition}
\alias{list_taxa_per_condition}
\title{Return a list of taxon_ids per condition}
\usage{
list_taxa_per_condition(ta, condition)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{condition}{The name of a variable in the samples table that contains a
categorical value.}
}
\value{
A list of taxon_id vectors.
}
\description{
This function returns a named list of unique taxon_ids per distinct value of a 
categorical column of the samples table.
}
