% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{make_tidyamplicons}
\alias{make_tidyamplicons}
\title{Create a tidyamplicons object from three tidy tables}
\usage{
make_tidyamplicons(
  samples,
  taxa,
  abundances,
  sample_name = sample,
  taxon_name = taxon
)
}
\description{
Create a tidyamplicons object from three tidy tables
}
