% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{make_tidytacos}
\alias{make_tidytacos}
\title{Create a tidytacos object from three tidy tables}
\usage{
make_tidytacos(
  samples,
  taxa,
  abundances,
  sample_name = sample,
  taxon_name = taxon
)
}
\description{
Create a tidytacos object from three tidy tables
}
