% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{markov_cluster}
\alias{markov_cluster}
\title{Performs Markov Clustering on a sparcc network matrix generated by \code{\link{sparcc_to_network}}. 
and returns a tibble with clusters and taxon identifiers. Optionally, the network can be visualized.}
\usage{
markov_cluster(network, min_n = 3, visualize = F)
}
\arguments{
\item{network}{sparcc network in matrix form.}

\item{min_n}{minimum number of taxa per cluster, smaller clusters are filtered out.}

\item{visualize}{wether to visualize the network clusters or not.}
}
\description{
Performs Markov Clustering on a sparcc network matrix generated by \code{\link{sparcc_to_network}}. 
and returns a tibble with clusters and taxon identifiers. Optionally, the network can be visualized.
}
