% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{mean_rel_abundances}
\alias{mean_rel_abundances}
\title{Get mean relative abundances of taxa in general or per condition}
\usage{
mean_rel_abundances(ta, condition = NULL)
}
\arguments{
\item{ta}{a tidyamplicons object}

\item{condition}{a string representing a categorical variable to compute the relative abundances in every option of the variable}
}
\description{
Returns tidy table with average relatively abundances of taxa, overall or per
condition.
}
\details{
Condition should be a categorical variable present in the samples table.
Supply condition as a string.
}
