% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{merge_tidyamplicons}
\alias{merge_tidyamplicons}
\title{Merge two tidyamplicons objects.}
\usage{
merge_tidyamplicons(ta1, ta2)
}
\arguments{
\item{ta1}{The first tidyamplicons object.}

\item{ta2}{The second tidyamplicons object.}
}
\description{
\code{merge_tidyamplicons} merges two tidyamplicons objects and returns one
single tidyamplicons object.
}
\details{
This function will merge two tidyamplicons objects into one. It is useful if
one wants to merge data obtained from different sequencing runs. Therefore,
this function requirers that both tidyamplicons objects contain a "run"
variable in their samples table, indicating their origin.
}
