% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{merge_tidytacos}
\alias{merge_tidytacos}
\title{Merge two tidytacos objects.}
\usage{
merge_tidytacos(ta1, ta2, taxon_identifier = sequence)
}
\arguments{
\item{ta1}{The first tidytacos object.}

\item{ta2}{The second tidytacos object.}
}
\description{
\code{merge_tidytacos} merges two tidytacos objects and returns one
single tidytacos object.
}
\details{
This function will merge two tidytacos objects into one. It is useful if
one wants to merge data obtained from different sequencing runs. Therefore,
this function requirers that both tidytacos objects contain a "run"
variable in their samples table, indicating their origin.
}
