% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{mutate_counts}
\alias{mutate_counts}
\title{Create extra variables in the abundances table}
\usage{
mutate_counts(ta, ...)
}
\arguments{
\item{ta}{a tidytacos object}
}
\description{
Create extra variables in the abundances table
}
