% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{mutate_taxa}
\alias{mutate_taxa}
\title{Create extra variables in the taxon table}
\usage{
mutate_taxa(ta, ...)
}
\arguments{
\item{ta}{a tidyamplicons object}
}
\description{
Create extra variables in the taxon table
}
