% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{numbers}
\alias{numbers}
\title{Return some descriptive numbers}
\usage{
numbers(ta)
}
\arguments{
\item{ta}{tidytacos object.}
}
\description{
\code{get_numbers} returns the number of samples, taxa and reads in a
tidytacos object.
}
\details{
This function returns the number of samples, taxa and reads in a
tidytacos object, stored in a named numeric vector. To print the numbers,
use \code{\link{report_numbers}} instead.
}
\examples{
# Initiate abundance matrix
x <- matrix(
 c(1500, 1300, 280, 356),
 ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x, taxa_are_columns = FALSE)

# Report numbers
numbers <- data \%>\% get_numbers()

}
