% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{pcoa_plot_ly}
\alias{pcoa_plot_ly}
\title{Return an interactive pcoa plot of the samples}
\usage{
pcoa_plot_ly(ta, x, samplenames = sample, palette = NULL, title = "PCOA plot")
}
\arguments{
\item{ta}{a tidytacos object}

\item{x}{a string, representing the column name used to color the sample groups on.}

\item{palette}{a vector of colors, used as the palette for coloring sample groups.}
}
\description{
Return an interactive pcoa plot of the samples
}
