% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{perform_adonis}
\alias{perform_adonis}
\title{Perform an adonis test}
\usage{
perform_adonis(ta, predictors, permutations = 999)
}
\description{
This function executes the adonis function of the vegan package and returns
the result.
}
\details{
Predictors should be a character vector. Samples where one of the predictors
is NA are removed.
}
