% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varia.R
\name{process_abundance_selection}
\alias{process_abundance_selection}
\title{Apply an abundance filtering to the taxon table}
\usage{
process_abundance_selection(ta)
}
\description{
DEPRECATED, see \code{\link{filter_abundances}}
}
\details{
Should only be used internally.
}
