% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varia.R
\name{process_sample_selection}
\alias{process_sample_selection}
\title{Apply a sample filtering to the taxon and abundance tables}
\usage{
process_sample_selection(ta)
}
\description{
DEPRECATED, see \code{\link{filter_samples}}
}
\details{
Should only be used internally.
}
