% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varia.R
\name{process_taxon_selection}
\alias{process_taxon_selection}
\title{Apply a taxon filtering to the abundance table}
\usage{
process_taxon_selection(ta)
}
\description{
DEPRECATED, see \code{\link{filter_taxa}}
}
\details{
Should only be used internally.
}
