% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{read_tidyamplicons}
\alias{read_tidyamplicons}
\title{Read community data written by tidyamplicons}
\usage{
read_tidyamplicons(
  din,
  samples = "samples.csv",
  taxa = "taxa.csv",
  abundances = "abundances.csv"
)
}
\description{
Read community data written by tidyamplicons
}
