% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{read_tidytacos}
\alias{read_tidytacos}
\title{Read community data written by tidytacos}
\usage{
read_tidytacos(
  din,
  samples = "samples.csv",
  taxa = "taxa.csv",
  counts = "counts.csv"
)
}
\arguments{
\item{din}{directory containing the a sample, taxa and counts table in csv format}

\item{samples}{the name of the samples table, defaults to samples.csv}

\item{taxa}{the name of the taxa table, defaults to taxa.csv}

\item{counts}{the name of the counts table, defaults to counts.csv}
}
\description{
Read community data written by tidytacos
}
