% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{rel_abundance_matrix}
\alias{rel_abundance_matrix}
\title{Return a relative abundance matrix}
\usage{
rel_abundance_matrix(ta, sample_name = sample, taxon_name = taxon)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{sample_name}{The name of the variable in the sample table to use as row
names (unquoted).}

\item{taxon_name}{The name of the variable in the taxon table to use as
column names (unquoted).}
}
\value{
A matrix with abundance values.
}
\description{
This function returns a relative abundance matrix; the rows are samples and
the column are taxa.
}
