% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_empty_samples}
\alias{remove_empty_samples}
\title{Removes empty samples from the tidyamplicons object}
\usage{
remove_empty_samples(ta)
}
\arguments{
\item{ta}{a tidyamplicons object}
}
\value{
the tidyamplicons object minus the empty samples
}
\description{
Removes empty samples from the tidyamplicons object
}
