% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{report_numbers}
\alias{report_numbers}
\title{Report numbers}
\usage{
report_numbers(ta)
}
\arguments{
\item{ta}{tidytacos object.}
}
\description{
\code{report_numbers} returns the number of samples, taxa and reads in a
tidytacos object.
}
\details{
This function prints the number of samples, taxa and reads in a tidytacos
object. To retrieve the numbers stored in named numeric vector, use
\code{\link{numbers}} instead.
}
\examples{
# Initiate abundance matrix
x <- matrix(
 c(1500, 1300, 280, 356),
 ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
                     taxa_are_columns = FALSE
                     )
# Report numbers
data \%>\%
 report_numbers()

}
