% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{sample_plot}
\alias{sample_plot}
\title{Return a visualization designed for a small number of samples}
\usage{
sample_plot(ta, sample = sample_id, n = 15, nrow = NULL)
}
\description{
Return a visualization designed for a small number of samples
}
