% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{select_abundances}
\alias{select_abundances}
\title{Retain or remove a set of abundance variables}
\usage{
select_abundances(ta, ...)
}
\arguments{
\item{ta}{a tidyamplicons object}
}
\description{
Retain or remove a set of abundance variables
}
