% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{select_counts}
\alias{select_counts}
\title{Retain or remove a set of count variables}
\usage{
select_counts(ta, ...)
}
\arguments{
\item{ta}{a tidytacos object}
}
\description{
Retain or remove a set of count variables
}
