% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{select_samples}
\alias{select_samples}
\title{Retain or remove a set of sample variables}
\usage{
select_samples(ta, ...)
}
\arguments{
\item{ta}{a tidyamplicons object}
}
\description{
Retain or remove a set of sample variables
}
