% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{select_taxa}
\alias{select_taxa}
\title{Retain or remove a set of taxon variables}
\usage{
select_taxa(ta, ...)
}
\arguments{
\item{ta}{a tidyamplicons object}
}
\description{
Retain or remove a set of taxon variables
}
