% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{sparcc}
\alias{sparcc}
\title{Perform network inference with SparCC on a tidytacos object, 
after rarefaction of the taxa. See \link[SpiecEasi]{sparcc}.}
\usage{
sparcc(
  ta,
  rarefact = 0.05,
  iter = 20,
  inner_iter = 10,
  th = 0.1,
  taxon_name = taxon,
  sample_name = sample
)
}
\arguments{
\item{ta}{a tidytacos object}

\item{rarefact}{Rarefaction degree: percentage of samples the taxon needs to be present in.}

\item{iter}{Number of iterations in the outer loop}

\item{inner_iter}{Number of iterations in the inner loop}

\item{th}{absolute value of correlations below this threshold are 
considered zero by the inner SparCC loop.}
}
\description{
Perform network inference with SparCC on a tidytacos object, 
after rarefaction of the taxa. See \link[SpiecEasi]{sparcc}.
}
