% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{sparcc_to_network}
\alias{sparcc_to_network}
\title{Transforms the output of \code{\link{sparcc}} into a matrix form 
for downstream clustering or heatplot visualization.}
\usage{
sparcc_to_network(sparcc.out, treshold = 0.1)
}
\arguments{
\item{treshold}{absolute value of correlations below this threshold are 
filtered out.}

\item{ta}{a tidytacos object}
}
\description{
Transforms the output of \code{\link{sparcc}} into a matrix form 
for downstream clustering or heatplot visualization.
}
