% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{tacoplot_ord}
\alias{tacoplot_ord}
\title{Return a pcoa plot of the samples}
\usage{
tacoplot_ord(ta, x = sample_id, palette = NULL, title = "PCOA plot")
}
\arguments{
\item{ta}{A tidytacos object.}

\item{x}{A string, representing the column name used to color the sample
groups on.}

\item{palette}{A vector of colors, used as the palette for coloring sample
groups.}
}
\description{
Return a pcoa plot of the samples
}
