% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{tacoplot_pcoa_ly}
\alias{tacoplot_pcoa_ly}
\title{Return an interactive pcoa plot of the samples}
\usage{
tacoplot_pcoa_ly(
  ta,
  x,
  samplenames = sample_id,
  palette = NULL,
  title = "PCOA plot"
)
}
\arguments{
\item{ta}{a tidytacos object}

\item{x}{a string, representing the column name used to color the sample groups on.}

\item{palette}{a vector of colors, used as the palette for coloring sample groups.}
}
\description{
Return an interactive pcoa plot of the samples
}
