% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{tacoplot_zoom}
\alias{tacoplot_zoom}
\title{Return a visualization designed for a small number of samples}
\usage{
tacoplot_zoom(ta, sample = sample_id, n = 15, nrow = NULL)
}
\description{
Return a visualization designed for a small number of samples
}
