% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{tacosum}
\alias{tacosum}
\title{Return some descriptive numbers}
\usage{
tacosum(ta)
}
\arguments{
\item{ta}{tidytacos object.}
}
\description{
\code{tacosum} returns the number of samples, taxa and reads in a
tidytacos object.
}
\examples{
# Initiate counts matrix
x <- matrix(
 c(1500, 1300, 280, 356),
 ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x, taxa_are_columns = FALSE)

# Report numbers
numbers <- data \%>\% tacosum()

}
