% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{taxon_counts_in_conditions}
\alias{taxon_counts_in_conditions}
\title{Get occurrences of taxa in general or in conditions}
\usage{
taxon_counts_in_conditions(ta, condition)
}
\description{
DEPRECATED, use \code{\link{occurrences}}
}
