% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{taxonlist_per_condition}
\alias{taxonlist_per_condition}
\title{Return a list of taxon IDs per condition}
\usage{
taxonlist_per_condition(ta, condition)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{condition}{The name of a variable in the sample table that contains a
categorical value.}
}
\value{
A list of taxon_id vectors.
}
\description{
This function returns a named list of unique taxon_ids per distinct value of
a categorical column of the samples table.
}
