% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyamplicons-package.R
\docType{package}
\name{tidyamplicons-package}
\alias{tidyamplicons}
\alias{tidyamplicons-package}
\title{tidyamplicons: Functions to Manipulate and Visualize Amplicon Abundance Data}
\description{
A toolkit for processing and visualizing amplicon abundance data, based on the tidyverse framework.
}
\author{
\strong{Maintainer}: Stijn Wittouck \email{stijn.wittouck@uantwerpen.be}

}
\keyword{internal}
