% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{trim_asvs}
\alias{trim_asvs}
\title{Trim all sequences}
\usage{
trim_asvs(ta, start, end)
}
\arguments{
\item{ta}{a tidytacos object}

\item{start}{index of where to start trimming}

\item{end}{index of where to stop trimming}
}
\description{
This function assumes that the sequence variable in the taxon table is called
"sequence".
}
