% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{update_id_names}
\alias{update_id_names}
\title{Update old tidyamplicons object to new one.}
\usage{
update_id_names(ta)
}
\arguments{
\item{ta}{Old tidyamplicons object.}
}
\description{
\code{update_id_names} updates an old tidyamplicons object to a new one.
}
\details{
This function will update a tidyamplicons object created prior to version
0.1.0 to a tidyamplicons object compatible with version 0.1.0.
}
