% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{update_tidytacos}
\alias{update_tidytacos}
\title{Update old tidytacos object to new one.}
\usage{
update_tidytacos(ta)
}
\arguments{
\item{ta}{Old tidytacos object.}
}
\description{
\code{update_tidytacos} updates an old tidytacos object to a new one.
}
\details{
This function will update a tidytacos object created prior to version
0.1.0 to a tidytacos object compatible with version 0.1.0.
}
