% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_lib_size}
\alias{add_lib_size}
\title{Add total abundance per sample}
\usage{
add_lib_size(ta, step = "current")
}
\arguments{
\item{ta}{tidytacos object.}
}
\description{
\code{add_lib_size} adds the total abundance per sample to the samples tibble
of a tidytacos object.
}
\details{
This function adds the total abundance per sample to the samples tibble of a
tidytacos object under the variable name total_abundance.
}
\examples{
# Initiate counts matrix
x <- matrix(
 c(1500, 1300, 280, 356),
 ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
                     taxa_are_columns = FALSE
                     )

# Add total counts
data <- data \%>\%
 add_lib_size()

}
