% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_taxa.R
\name{add_mean_rel_abundances}
\alias{add_mean_rel_abundances}
\title{Add average relative abundances}
\usage{
add_mean_rel_abundances(ta, condition = NULL, wilcox = F)
}
\arguments{
\item{ta}{A tidyamplicons object}

\item{condition}{A condition variable (character)}

\item{wilcox}{Should a wilcoxon rank sum tests be performed}
}
\value{
A tidyamplicons object
}
\description{
This function adds mean relative abundance values for each taxon to the taxa
table, overall or per sample group.
}
\details{
If `condition` is specified, the mean relative abundances will be calculated
separately for each group defined by the condition variable. This variable
should be present in the sample table.

If `wilcox` is TRUE, a wilcoxon rank sum test is performed to test
differential abundance of taxa between groups of samples.
}
