% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_abundances.R
\name{add_rel_abundance}
\alias{add_rel_abundance}
\title{Add relative abundance to abundance table}
\usage{
add_rel_abundance(ta)
}
\arguments{
\item{ta}{tidytacos object.}
}
\description{
\code{add_rel_abundance} adds relative abundance to the abundance table of a
tidytacos object.
}
\details{
This function adds the relative abundance per sample to the abundance table
of a tidytacos object under the variable name "rel_abundance".
}
\examples{
# Initiate abundance matrix
x <- matrix(
 c(1500, 1300, 280, 356),
 ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
                     taxa_are_columns = FALSE)

# Add relative abundance
data <- data \%>\%
 add_rel_abundance()

}
