% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_sample_clustered}
\alias{add_sample_clustered}
\title{Add clustered sample order}
\usage{
add_sample_clustered(ta)
}
\arguments{
\item{ta}{tidytacos object.}
}
\description{
\code{add_sample_clustered} adds a new variable defining a sample order based
on similarity after clustering to the samples tibble of a tidytacos
object.
}
\details{
This function calculates the Bray-Curtis distance between samples followed by
hierarchical average linkage clustering of samples. It will then add a new
factor variable "samples_clustered" to the samples tibble of a tidytacos
object. This function is extremely useful if one wants to plot similar
samples together.
}
\examples{
# Initiate abundance matrix
x <- matrix(
 c(1500, 1300, 280, 356),
 ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
                     taxa_are_columns = FALSE
                     )

# Add total abundance
data <- data \%>\%
 add_sample_clustered()

}
