% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_spike_ratio}
\alias{add_spike_ratio}
\title{Add spike ratio}
\usage{
add_spike_ratio(ta, spike_taxon)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{spike_taxon}{The taxon_id of the spike.}
}
\description{
\code{add_spike_ratio} adds a new variable showing the ratio total abundance
to spike abundance to the samples tibble of a tidytacos object.
}
\details{
This function calculates the spike ratio defined as the total sample
abundance to the spike abundance and adds this as a new variable
"spike_ratio" to the samples tibble of a tidytacos object. This function
is useful if a DNA spike was added prior to sequencing and is based on the
method described by
\href{https://doi.org/10.1016/j.soilbio.2016.02.003}{Smets et al., 2016}.

Credits to Wenke Smets for the idea of spiking samples prior to 16S
sequencing and the initial implementation of this function.
}
\examples{
# Initiate abundance matrix
x <- matrix(
 c(1500, 1300, 280, 356),
 ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
                     taxa_are_columns = FALSE
                     )

# Add total abundance
data <- data \%>\%
 add_spike_ratio(spike_taxon = "t1")
}
