% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_taxa.R
\name{add_taxon_name}
\alias{add_taxon_name}
\title{Create sensible names for the taxa and add to taxon table}
\usage{
add_taxon_name(ta, method = "total_rel_abundance", include_species = F)
}
\description{
Create sensible names for the taxa and add to taxon table
}
