% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_taxa.R
\name{add_taxon_name_color}
\alias{add_taxon_name_color}
\title{Create taxon names suitable for visualization with color. A rank can be supplied to aggregate colors higher than the current rank.}
\usage{
add_taxon_name_color(
  ta,
  method = "total_rel_abundance",
  n = 12,
  samples = NULL,
  taxa = NULL,
  rank = NULL
)
}
\arguments{
\item{ta}{a tidyamplicons object}

\item{method}{the method on which to arrange the taxon names.}

\item{n}{integer denoting the amount of most abundant taxa to display. Capacity at 12.}

\item{samples}{optional vector of sample_id's of interest}

\item{taxa}{optional vector of taxon_id's of interest}
}
\description{
Create taxon names suitable for visualization with color. A rank can be supplied to aggregate colors higher than the current rank.
}
