% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_taxa.R
\name{add_total_rel_abundance}
\alias{add_total_rel_abundance}
\title{Add the total relative abundance of taxa to the taxon table}
\usage{
add_total_rel_abundance(ta)
}
\description{
Add the total relative abundance of taxa to the taxon table
}
