% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{as_tidyamplicons}
\alias{as_tidyamplicons}
\title{Convert phyloseq object to tidyamplicons object}
\usage{
as_tidyamplicons(ps)
}
\arguments{
\item{ps}{Phyloseq object.}
}
\description{
\code{as_tidyamplicons} returns a tidyamplicons object given a phyloseq
object.
}
\details{
This function will convert a phyloseq object into a tidyamplicons object. To
convert from a tidyamplicons object to a phyloseq object use
\code{\link{as_phyloseq}}.
}
