% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_test_ta}
\alias{create_test_ta}
\title{Create a tidyamplicons object for testing/example purporses}
\usage{
create_test_ta()
}
\value{
a small tidyamplicons object
}
\description{
Create a tidyamplicons object for testing/example purporses
}
