% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{perform_adonis}
\alias{perform_adonis}
\title{Perform an adonis test}
\usage{
perform_adonis(ta, predictors, permutations = 999)
}
\arguments{
\item{ta}{A tidyamplicons object.}

\item{predictors}{A character vector with predictors to include in the model.}

\item{permutations}{The number of permutations (more permutations takes
longer but gives a more accurate p-value).}
}
\value{
An object of class "adonis" (see \link[vegan]{adonis}).
}
\description{
This function executes the \link[vegan]{adonis} function of the vegan package
and returns the result.
}
\details{
Samples where one or more predictors are NA are removed.
}
