% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{read_tidyamplicons}
\alias{read_tidyamplicons}
\title{Read community data written by tidyamplicons}
\usage{
read_tidyamplicons(
  din,
  samples = "samples.csv",
  taxa = "taxa.csv",
  abundances = "abundances.csv"
)
}
\arguments{
\item{din}{directory containing the a sample, taxa and abundances table in csv format}

\item{samples}{the name of the samples table, defaults to samples.csv}

\item{taxa}{the name of the taxa table, defaults to taxa.csv}

\item{abundances}{the name of the abundances table, defaults to abundances.csv}
}
\description{
Read community data written by tidyamplicons
}
