% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_empty_samples}
\alias{remove_empty_samples}
\title{Removes empty samples from the tidytacos object}
\usage{
remove_empty_samples(ta)
}
\arguments{
\item{ta}{a tidytacos object}
}
\value{
the tidytacos object minus the empty samples
}
\description{
Removes empty samples from the tidytacos object
}
