% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{tacoplot_stack_ly}
\alias{tacoplot_stack_ly}
\title{Return an interactive bar plot of the samples}
\usage{
tacoplot_stack_ly(ta, n = 12, x = sample_clustered)
}
\arguments{
\item{ta}{a tidytacos object}

\item{n}{an integer, representing the amount of colors used to depict different taxa}

\item{x}{a string, representing the column name used to label and cluster the samples on.}
}
\description{
Return an interactive bar plot of the samples
}
