% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{update_id_names}
\alias{update_id_names}
\title{Rename the "sample" and "taxon" columns to "sample_id" and "taxon_id"}
\usage{
update_id_names(ta)
}
\arguments{
\item{ta}{a tidyamplicons object}
}
\description{
Rename the "sample" and "taxon" columns to "sample_id" and "taxon_id"
}
