% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{write_tidytacos}
\alias{write_tidytacos}
\title{Write community data in tidytacos format}
\usage{
write_tidytacos(ta, dout)
}
\arguments{
\item{ta}{a tidytacos object}

\item{dout}{the directory to store the three tidytacos tables in}
}
\description{
Write community data in tidytacos format
}
