% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{write_tidyamplicons}
\alias{write_tidyamplicons}
\title{Write community data in tidyamplicons format}
\usage{
write_tidyamplicons(ta, dout)
}
\arguments{
\item{ta}{a tidyamplicons object}

\item{dout}{the directory to store the three tidyamplicons tables in}
}
\description{
Write community data in tidyamplicons format
}
